refactoring of metrics and comparisons
