+++ climpred/__init__.py
@@ -4,5 +4,7 @@ from . import loadutils
 from . import relative_entropy
 from . import graphics
 from . import bootstrap
+from . import metrics
+from . import comparisons
 # eventually will *only* import these
 from .classes import PerfectModelEnsemble, ReferenceEnsemble
