+++ climpred/comparisons.py
@@ -1,0 +1,209 @@
+from .stats import _get_dims
+from .prediction import _stack_to_supervector, _drop_members
+import xarray as xr
+
+
+def _get_comparison_function(comparison):
+    """Converts a string comparison entry from the user into an actual
+       function for the package to interpret.
+
+    PERFECT MODEL:
+    m2m: Compare all members to all other members.
+    m2c: Compare all members to the control.
+    m2e: Compare all members to the ensemble mean.
+    e2c: Compare the ensemble mean to the control.
+
+    REFERENCE:
+    e2r: Compare the ensemble mean to the reference.
+    m2r: Compare each ensemble member to the reference.
+
+    Args:
+        comparison (str): name of comparison.
+
+    Returns:
+        comparison (function): comparison function.
+
+    """
+    if comparison == 'm2m':
+        comparison = '_m2m'
+    elif comparison == 'm2c':
+        comparison = '_m2c'
+    elif comparison == 'm2e':
+        comparison = '_m2e'
+    elif comparison == 'e2c':
+        comparison = '_e2c'
+    elif comparison == 'e2r':
+        comparison = '_e2r'
+    elif comparison == 'm2r':
+        comparison = '_m2r'
+    else:
+        raise ValueError("""Please supply a comparison from the following list:
+            'm2m'
+            'm2c'
+            'm2e'
+            'e2c'
+            'e2r'
+            'm2r'
+            """)
+    return eval(comparison)
+
+
+def _m2m(ds, supervector_dim='svd'):
+    """
+    Create two supervectors to compare all members to all other members in
+    turn.
+
+    Args:
+        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
+                            dimension.
+        supervector_dim (str): name of new supervector dimension.
+                               Default: 'svd'
+
+    Returns:
+        forecast (xarray object): forecast.
+        reference (xarray object): reference.
+
+    """
+    reference_list = []
+    forecast_list = []
+    for m in ds.member.values:
+        # drop the member being reference
+        ds_reduced = _drop_members(ds, rmd_member=[m])
+        reference = ds.sel(member=m)
+        for m2 in ds_reduced.member:
+            for i in ds.initialization:
+                reference_list.append(reference.sel(initialization=i))
+                forecast_list.append(
+                    ds_reduced.sel(member=m2, initialization=i))
+    reference = xr.concat(reference_list, supervector_dim)
+    forecast = xr.concat(forecast_list, supervector_dim)
+    return forecast, reference
+
+
+def _m2e(ds, supervector_dim='svd'):
+    """
+    Create two supervectors to compare all members to ensemble mean.
+
+    Args:
+        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
+                            dimension.
+        supervector_dim (str): name of new supervector dimension.
+                               Default: 'svd'
+
+    Returns:
+        forecast (xarray object): forecast.
+        reference (xarray object): reference.
+
+    """
+    reference = ds.mean('member')
+    forecast, reference = xr.broadcast(ds, reference)
+    forecast = _stack_to_supervector(forecast, new_dim=supervector_dim)
+    reference = _stack_to_supervector(reference, new_dim=supervector_dim)
+    return forecast, reference
+
+
+def _m2c(ds, supervector_dim='svd', control_member=[0]):
+    """
+    Create two supervectors to compare all members to control.
+
+    Args:
+        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
+                            dimension.
+        supervector_dim (str): name of new supervector dimension.
+                               Default: 'svd'
+        control_member: list of the one integer member serving as
+                        reference. Default 0
+
+    Returns:
+        forecast (xarray object): forecast.
+        reference (xarray object): reference.
+
+    """
+    reference = ds.isel(member=control_member).squeeze()
+    # drop the member being reference
+    ds_dropped = _drop_members(ds, rmd_member=ds.member.values[control_member])
+    forecast, reference = xr.broadcast(ds_dropped, reference)
+    forecast = _stack_to_supervector(forecast, new_dim=supervector_dim)
+    reference = _stack_to_supervector(reference, new_dim=supervector_dim)
+    return forecast, reference
+
+    return forecast, reference
+
+
+def _e2c(ds, supervector_dim='svd', control_member=[0]):
+    """
+    Create two supervectors to compare ensemble mean to control.
+
+    Args:
+        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
+                            dimension.
+        supervector_dim (str): name of new supervector dimension.
+                               Default: 'svd'
+        control_member: list of the one integer member serving as
+                        reference. Default 0
+
+    Returns:
+        forecast (xarray object): forecast.
+        reference (xarray object): reference.
+    """
+    reference = ds.isel(member=control_member).squeeze()
+    reference = reference.rename({'initialization': supervector_dim})
+    # drop the member being reference
+    ds = _drop_members(ds, rmd_member=[ds.member.values[control_member]])
+    forecast = ds.mean('member')
+    forecast = forecast.rename({'initialization': supervector_dim})
+    return forecast, reference
+
+
+def _e2r(ds, reference):
+    """
+    For a reference-based decadal prediction ensemble. This compares the
+    ensemble mean prediction to the reference (hindcast, simulation,
+    observations).
+
+    Args:
+        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
+                            dimension.
+        reference (xarray object): reference xr.Dataset/xr.DataArray.
+
+    Returns:
+        forecast (xarray object): forecast.
+        reference (xarray object): reference.
+
+    """
+    if 'member' in _get_dims(ds):
+        print("Taking ensemble mean...")
+        forecast = ds.mean('member')
+    else:
+        forecast = ds
+    return forecast, reference
+
+
+def _m2r(ds, reference):
+    """
+    For a reference-based decadal prediction ensemble. This compares each
+    member individually to the reference (hindcast, simulation,
+    observations).
+
+    Args:
+        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
+                            dimension.
+        reference (xarray object): reference xr.Dataset/xr.DataArray.
+
+    Returns:
+        forecast (xarray object): forecast.
+        reference (xarray object): reference.
+
+    """
+    # check that this contains more than one member
+    if ('member' not in _get_dims(ds)) or (ds.member.size == 1):
+        raise ValueError("""Please supply a decadal prediction ensemble with
+            more than one member. You might have input the ensemble mean here
+            although asking for a member-to-reference comparison.""")
+    else:
+        forecast = ds
+    reference = reference.expand_dims('member')
+    nMember = forecast.member.size
+    reference = reference.isel(member=[0] * nMember)
+    reference['member'] = forecast['member']
+    return forecast, reference
