+++ climpred/prediction.py
@@ -1,9 +1,7 @@
 """Objects dealing with decadal prediction metrics."""
-import types
 import warnings
 
 import cftime
-import dask
 import numpy as np
 import xarray as xr
 
@@ -14,6 +12,10 @@ from xskillscore import pearson_r_p_value
 from xskillscore import rmse as _rmse
 
 from .stats import _check_xarray, _get_dims, z_significance
+from .comparisons import (_get_comparison_function, _m2m, _m2c,
+                          _m2e, _e2c, _e2r, _m2r)
+from .metrics import (_get_metric_function, _nmae, _nrmse,
+                      _nmse, _ppp, _uacc)
 
 
 # -------------------------------------------- #
@@ -190,8 +192,8 @@ def _select_members_ensembles(ds, m=None, i=None):
 def _stack_to_supervector(ds,
                           new_dim='svd',
                           stacked_dims=('initialization', 'member')):
-    """Stack all stacked_dims (likely initialization and member) dimensions into one
-    supervector dimension to perform metric over.
+    """Stack all stacked_dims (likely initialization and member) dimensions
+    into one supervector dimension to perform metric over.
 
     Args:
         ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
@@ -207,465 +209,10 @@ def _stack_to_supervector(ds,
 
 
 # --------------------------------------------#
-# COMPARISONS
-# --------------------------------------------#
-def _get_comparison_function(comparison):
-    """Converts a string comparison entry from the user into an actual
-       function for the package to interpret.
-
-    PERFECT MODEL:
-    m2m: Compare all members to all other members.
-    m2c: Compare all members to the control.
-    m2e: Compare all members to the ensemble mean.
-    e2c: Compare the ensemble mean to the control.
-
-    REFERENCE:
-    e2r: Compare the ensemble mean to the reference.
-    m2r: Compare each ensemble member to the reference.
-
-    Args:
-        comparison (str): name of comparison.
-
-    Returns:
-        comparison (function): comparison function.
-
-    """
-    if comparison == 'm2m':
-        comparison = '_m2m'
-    elif comparison == 'm2c':
-        comparison = '_m2c'
-    elif comparison == 'm2e':
-        comparison = '_m2e'
-    elif comparison == 'e2c':
-        comparison = '_e2c'
-    elif comparison == 'e2r':
-        comparison = '_e2r'
-    elif comparison == 'm2r':
-        comparison = '_m2r'
-    else:
-        raise ValueError("""Please supply a comparison from the following list:
-            'm2m'
-            'm2c'
-            'm2e'
-            'e2c'
-            'e2r'
-            'm2r'
-            """)
-    return eval(comparison)
-
-
-def _m2m(ds, supervector_dim='svd'):
-    """
-    Create two supervectors to compare all members to all other members in turn.
-
-    Args:
-        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
-                            dimension.
-        supervector_dim (str): name of new supervector dimension. Default: 'svd'
-
-    Returns:
-        forecast (xarray object): forecast.
-        reference (xarray object): reference.
-
-    """
-    reference_list = []
-    forecast_list = []
-    for m in ds.member.values:
-        # drop the member being reference
-        ds_reduced = _drop_members(ds, rmd_member=[m])
-        reference = ds.sel(member=m)
-        for m2 in ds_reduced.member:
-            for i in ds.initialization:
-                reference_list.append(reference.sel(initialization=i))
-                forecast_list.append(
-                    ds_reduced.sel(member=m2, initialization=i))
-    reference = xr.concat(reference_list, supervector_dim)
-    forecast = xr.concat(forecast_list, supervector_dim)
-    return forecast, reference
-
-
-def _m2e(ds, supervector_dim='svd'):
-    """
-    Create two supervectors to compare all members to ensemble mean.
-
-    Args:
-        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
-                            dimension.
-        supervector_dim (str): name of new supervector dimension. Default: 'svd'
-
-    Returns:
-        forecast (xarray object): forecast.
-        reference (xarray object): reference.
-
-    """
-    reference = ds.mean('member')
-    forecast, reference = xr.broadcast(ds, reference)
-    forecast = _stack_to_supervector(forecast, new_dim=supervector_dim)
-    reference = _stack_to_supervector(reference, new_dim=supervector_dim)
-    return forecast, reference
-
-
-def _m2c(ds, supervector_dim='svd', control_member=[0]):
-    """
-    Create two supervectors to compare all members to control.
-
-    Args:
-        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
-                            dimension.
-        supervector_dim (str): name of new supervector dimension. Default: 'svd'
-        control_member: list of the one integer member serving as
-                        reference. Default 0
-
-    Returns:
-        forecast (xarray object): forecast.
-        reference (xarray object): reference.
-
-    """
-    reference = ds.isel(member=control_member).squeeze()
-    # drop the member being reference
-    ds_dropped = _drop_members(ds, rmd_member=ds.member.values[control_member])
-    forecast, reference = xr.broadcast(ds_dropped, reference)
-    forecast = _stack_to_supervector(forecast, new_dim=supervector_dim)
-    reference = _stack_to_supervector(reference, new_dim=supervector_dim)
-    return forecast, reference
-
-    return forecast, reference
-
-
-def _e2c(ds, supervector_dim='svd', control_member=[0]):
-    """
-    Create two supervectors to compare ensemble mean to control.
-
-    Args:
-        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
-                            dimension.
-        supervector_dim (str): name of new supervector dimension. Default: 'svd'
-        control_member: list of the one integer member serving as
-                        reference. Default 0
-
-    Returns:
-        forecast (xarray object): forecast.
-        reference (xarray object): reference.
-    """
-    reference = ds.isel(member=control_member).squeeze()
-    reference = reference.rename({'initialization': supervector_dim})
-    # drop the member being reference
-    ds = _drop_members(ds, rmd_member=[ds.member.values[control_member]])
-    forecast = ds.mean('member')
-    forecast = forecast.rename({'initialization': supervector_dim})
-    return forecast, reference
-
-
-def _e2r(ds, reference):
-    """
-    For a reference-based decadal prediction ensemble. This compares the
-    ensemble mean prediction to the reference (hindcast, simulation,
-    observations).
-
-    Args:
-        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
-                            dimension.
-        reference (xarray object): reference xr.Dataset/xr.DataArray.
-
-    Returns:
-        forecast (xarray object): forecast.
-        reference (xarray object): reference.
-
-    """
-    if 'member' in _get_dims(ds):
-        print("Taking ensemble mean...")
-        forecast = ds.mean('member')
-    else:
-        forecast = ds
-    return forecast, reference
-
-
-def _m2r(ds, reference):
-    """
-    For a reference-based decadal prediction ensemble. This compares each
-    member individually to the reference (hindcast, simulation,
-    observations).
-
-    Args:
-        ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
-                            dimension.
-        reference (xarray object): reference xr.Dataset/xr.DataArray.
-
-    Returns:
-        forecast (xarray object): forecast.
-        reference (xarray object): reference.
-
-    """
-    # check that this contains more than one member
-    if ('member' not in _get_dims(ds)) or (ds.member.size == 1):
-        raise ValueError("""Please supply a decadal prediction ensemble with
-            more than one member. You might have input the ensemble mean here
-            although asking for a member-to-reference comparison.""")
-    else:
-        forecast = ds
-    reference = reference.expand_dims('member')
-    nMember = forecast.member.size
-    reference = reference.isel(member=[0] * nMember)
-    reference['member'] = forecast['member']
-    return forecast, reference
-
-
-# --------------------------------------------#
-# METRICS
-# Metrics for computing predictability.
-# --------------------------------------------#
-def _get_metric_function(metric):
-    """
-    This allows the user to submit a string representing the desired function
-    to anything that takes a metric.
-
-    Currently compatable with functions:
-    * compute_persistence()
-    * compute_perfect_model()
-    * compute_reference()
-
-    Currently compatable with metrics:
-    * pearson_r
-    * rmse
-    * mae
-    * mse
-    * nrmse
-    * nmae
-    * nmse
-    * msss
-    * uacc
-
-    Args:
-        metric (str): name of metric.
-
-    Returns:
-        metric (function): function object of the metric.
-
-    Raises:
-        ValueError: if metric not implemented.
-
-    """
-    # catches issues with wrappers, etc. that actually submit the
-    # proper underscore function
-    if type(metric) == types.FunctionType:
-        return metric
-    else:
-        pearson = ['pr', 'pearsonr', 'pearson_r']
-        if metric.lower() in pearson:
-            metric = '_pearson_r'
-        elif metric.lower() == 'rmse':
-            metric = '_rmse'
-        elif metric.lower() == 'mae':
-            metric = '_mae'
-        elif metric.lower() == 'mse':
-            metric = '_mse'
-        elif metric.lower() == 'nrmse':
-            metric = '_nrmse'
-        elif metric.lower() in ['nev', 'nmse']:
-            metric = '_nmse'
-        elif metric.lower() in ['ppp', 'msss']:
-            metric = '_ppp'
-        elif metric.lower() == 'nmae':
-            metric = '_nmae'
-        elif metric.lower() == 'uacc':
-            metric = '_uacc'
-        else:
-            raise ValueError("""Please supply a metric from the following list:
-                'pearson_r'
-                'rmse'
-                'mae'
-                'mse'
-                'nrmse'
-                'nev'
-                'nmse'
-                'ppp'
-                'msss'
-                'nmae'
-                'uacc'
-                """)
-        return eval(metric)
-
-
-# TODO: Do we need wrappers or should we rather create wrappers for skill score
-#       as used in a specific paper: def Seferian2018(ds, control):
-#       return PM_compute(ds, control, metric=_ppp, comparison=_m2e)
-def _ppp(ds, control, comparison, running=None, reference_period=None):
-    """Prognostic Potential Predictability (PPP) metric.
-
-    .. math:: PPP = 1 - \frac{MSE}{ \sigma_{control} \cdot fac}
-    Perfect forecast: 1
-    Climatology forecast: 0
-
-    References:
-      * Griffies, S. M., and K. Bryan. “A Predictability Study of Simulated
-        North Atlantic Multidecadal Variability.” Climate Dynamics 13, no. 7–8
-        (August 1, 1997): 459–87. https://doi.org/10/ch4kc4.
-      * Pohlmann, Holger, Michael Botzet, Mojib Latif, Andreas Roesch, Martin
-        Wild, and Peter Tschuck. “Estimating the Decadal Predictability of a
-        Coupled AOGCM.” Journal of Climate 17, no. 22 (November 1, 2004):
-        4463–72. https://doi.org/10/d2qf62.
-      * Bushuk, Mitchell, Rym Msadek, Michael Winton, Gabriel Vecchi, Xiaosong
-        Yang, Anthony Rosati, and Rich Gudgel. “Regional Arctic Sea–Ice
-        Prediction: Potential versus Operational Seasonal Forecast Skill.
-        Climate Dynamics, June 9, 2018. https://doi.org/10/gd7hfq.
-
-    Args:
-        ds (xarray object): xr.Dataset/xr.DataArray with member dimension.
-        control (xarray object): xr.Dataset/xr.DataArray of control simulation.
-        comparison (function): comparison function.
-        running (int): smoothing of control. Default: None (no smoothing).
-        reference_period (str): see _control_for_reference_period.
-
-    Returns:
-        ppp_skill (xarray object): skill of PPP.
-
-    """
-    supervector_dim = 'svd'
-    forecast, reference = comparison(ds, supervector_dim)
-    mse_skill = _mse(forecast, reference, dim=supervector_dim)
-    var = _get_variance(
-        control, time_length=running, reference_period=reference_period)
-    fac = _get_norm_factor(comparison)
-    ppp_skill = 1 - mse_skill / var / fac
-    return ppp_skill
-
-
-def _nrmse(ds, control, comparison, running=None, reference_period=None):
-    """Normalized Root Mean Square Error (NRMSE) metric.
-
-    .. math:: NRMSE = \frac{RMSE}{\sigma_{control} \cdot \sqrt{fac}
-                    = sqrt{ \frac{MSE}{ \sigma^2_{control} \cdot fac} }
-
-    Perfect forecast: 0
-    Climatology forecast: 1
-
-
-    References:
-      * Bushuk, Mitchell, Rym Msadek, Michael Winton, Gabriel Vecchi, Xiaosong
-        Yang, Anthony Rosati, and Rich Gudgel. “Regional Arctic Sea–Ice
-        Prediction: Potential versus Operational Seasonal Forecast Skill.”
-        Climate Dynamics, June 9, 2018. https://doi.org/10/gd7hfq.
-      * Hawkins, Ed, Steffen Tietsche, Jonathan J. Day, Nathanael Melia, Keith
-        Haines, and Sarah Keeley. “Aspects of Designing and Evaluating
-        Seasonal-to-Interannual Arctic Sea-Ice Prediction Systems.” Quarterly
-        Journal of the Royal Meteorological Society 142, no. 695
-        (January 1, 2016): 672–83. https://doi.org/10/gfb3pn.
-
-    Args:
-        ds (xarray object): xr.Dataset/xr.DataArray with member dimension.
-        control (xarray object): xr.Dataset/xr.DataArray of control simulation.
-        comparison (function): comparison function.
-        running (int): smoothing of control. Default: None (no smoothing).
-        reference_period (str): see _control_for_reference_period.
-
-    Returns:
-        nrmse_skill (xarray object): skill of NRMSE.
-
-    """
-    supervector_dim = 'svd'
-    forecast, reference = comparison(ds, supervector_dim)
-    rmse_skill = _rmse(forecast, reference, dim=supervector_dim)
-    var = _get_variance(
-        control, time_length=running, reference_period=reference_period)
-    fac = _get_norm_factor(comparison)
-    nrmse_skill = 1 - rmse_skill / np.sqrt(var) / np.sqrt(fac)
-    return nrmse_skill
-
-
-def _nmse(ds, control, comparison, running=None, reference_period=None):
-    """
-    Normalized MSE (NMSE) = Normalized Ensemble Variance (NEV) metric.
-
-
-    .. math:: NMSE = NEV = frac{MSE}{\sigma^2_{control} \cdot fac}
-
-    Perfect forecast: 0
-    Climatology forecast: 1
-
-    Reference:
-    * Griffies, S. M., and K. Bryan. “A Predictability Study of Simulated North
-      Atlantic Multidecadal Variability.” Climate Dynamics 13, no. 7–8
-      (August 1, 1997): 459–87. https://doi.org/10/ch4kc4.
-
-    Args:
-        ds (xarray object): xr.Dataset/xr.DataArray with member dimension.
-        control (xarray object): xr.Dataset/xr.DataArray of control simulation.
-        comparison (function): comparison function.
-        running (int): smoothing of control. Default: None (no smoothing).
-        reference_period (str): see _control_for_reference_period.
-
-    Returns:
-        nmse_skill (xarray object): skill of NMSE.
-    """
-    supervector_dim = 'svd'
-    forecast, reference = comparison(ds, supervector_dim)
-    mse_skill = _mse(forecast, reference, dim=supervector_dim)
-    var = _get_variance(
-        control, time_length=running, reference_period=reference_period)
-    fac = _get_norm_factor(comparison)
-    nmse_skill = 1 - mse_skill / var / fac
-    return nmse_skill
-
-
-def _nmae(ds, control, comparison, running=None, reference_period=None):
-    """
-    Normalized Ensemble Mean Absolute Error metric.
-
-    Formula
-    -------
-    NMAE-SS = 1 - mse / var
-
-    Reference
-    ---------
-    - Griffies, S. M., and K. Bryan. “A Predictability Study of Simulated North
-      Atlantic Multidecadal Variability.” Climate Dynamics 13, no. 7–8
-      (August 1, 1997): 459–87. https://doi.org/10/ch4kc4.
-
-      NOTE: NMSE = - 1 - NEV
-    """
-    supervector_dim = 'svd'
-    fct, truth = comparison(ds, supervector_dim)
-    mse_skill = _mse(fct, truth, dim=supervector_dim)
-    var = _get_variance(
-        control, time_length=running, reference_period=reference_period)
-    fac = _get_norm_factor(comparison)
-    nmse_skill = 1 - mse_skill / var / fac
-    return nmse_skill
-
-
-def _uacc(forecast, reference, control, running=None, reference_period=None):
-    """
-    Unbiased ACC (uACC) metric.
-
-    .. math::
-        uACC = \sqrt{PPP} = \sqrt{MSSS}
-
-    Reference
-    * Bushuk, Mitchell, Rym Msadek, Michael Winton, Gabriel Vecchi, Xiaosong
-      Yang, Anthony Rosati, and Rich Gudgel. “Regional Arctic Sea–Ice
-      Prediction: Potential versus Operational Seasonal Forecast Skill.
-      Climate Dynamics, June 9, 2018. https://doi.org/10/gd7hfq.
-
-    Args:
-        ds (xarray object): xr.Dataset/xr.DataArray with member dimension.
-        control (xarray object): xr.Dataset/xr.DataArray of control simulation.
-        comparison (function): comparison function.
-        running (int): smoothing of control. Default: None (no smoothing).
-        reference_period (str): see _control_for_reference_period.
-
-    Returns:
-        uacc_skill (xarray object): skill of uACC
-    """
-    return np.sqrt(
-        _ppp(forecast, reference, control, running, reference_period))
-
-
-# --------------------------------------------#
 # COMPUTE PREDICTABILITY/FORECASTS
 # Highest-level features for computing
 # predictability.
 # --------------------------------------------#
-
 def compute_perfect_model(ds,
                           control,
                           metric='pearson_r',
@@ -847,7 +394,9 @@ def compute_persistence_pm(ds, control, nlags, metric='pearson_r',
             'mae'""")
 
     init_years = ds['initialization'].values
-    if isinstance(ds.time.values[0], cftime._cftime.DatetimeProlepticGregorian) or isinstance(ds.time.values[0], np.datetime64):
+    if isinstance(ds.time.values[0],
+                  cftime._cftime.DatetimeProlepticGregorian) \
+            or isinstance(ds.time.values[0], np.datetime64):
         init_cftimes = []
         for year in init_years:
             init_cftimes.append(control.sel(
